refactor: extract image app context helpers
